Refactor: weatherclinetTest 전략 별 분리
